SpecSAT: simple benchmarking
By now, the generator and solver can be used, in a sequential setting.
Parallel solving is not done, nor a good benchmark selection.
Furthermore, the output is not yet logged to a file, not is the relevant
output presented in a meaningful way. Finally, the evaluation of the
data still has to be implemented.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>