benchmarks: restrict target
To have meaningful benchmarks for the sequential and parallel run, we
need to make sure that we have good candidate formulas in the set. As
parallel and sequential search behaves differently - biggest difference
being that no sequential simplification is run before the parallel
search - the benchmarks also have to be different.

Hence, this change implements a restriction on benchmarks to run.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>