validate: check expected conflicts
... so that we can be sure the expected execution would actually
happen during measurement. The number of conflicts is used for
this process. Failures are propagated.

For now, only use correct expectations for the single threaded
use case, as there is no simple mechanism to calculate the
number of conflicts for the parallel solving algorithm at this
point - even though it is deterministic.
that is used during the lite mode.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>