exec,dir: run from local dir by default
When running SpecSat, the base directory should be the current working
directory. Hence, modify the behavior accordingly.

However, when running in a distributed environment with shared network
file systems, or when using a batch compute system with a temporary
working directory, support to run from withing the install directory
by default. This allows to run SpecSAT on multiple systems easily.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>