solver: use deterministic math
MergeSat now offers a flag to switch between using the math libaries
pow function, or an implementation that ships with MergeSat. Using
the internal implementation results in a more reliable build. Hence,
we use this.

As the search in the solver changes with that switch as well, the
expected conflicts per benchmark need to be adapted.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>