parallel,limit: use conflict limits
The recent MergeSat solver supports a conflict limit in the
sequential and parallel mode. This conflict limit is enforced
in the first solver used inside the solver. As the conflicts
of this solver need to be below the conflict SUM when solving
the formula in parallel, use this as an upper bound to further
limit the runtime of SpecSAT on platforms where it is not
deterministic.