input: support CNFs from storage
To work around differences in the generator across platforms, we
also support reading CNF file from an existing directory. This way,
the CNFs do not have to be generated. The hashsum of the formulas
will still be checked, so that we know whether we work on the
expected input.

This feature helps to use the benchmark across more platforms.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>